Integrate Invoice Generation, Restore Edit Invoice Functionality (#14)
* integrate invoice generation

* bring receipt url to payment table

* license updates

* checkpt - checkout needs testing

* fix missing payment form

* subscription status updates, setup intent processing is now idempotent

* setup intent success

* tests passing again

* add migration files

* lint passes for multi-type purchase session

* pdf generation works, purchase sessions for invoices work

* unit test tweaks

* invoices table working

* hello world for invoices page

* connect select

* checkpt

* invoice side navigation

* customer details billing tab

* invoice reminder email working well

* customer facing invoice page and stripe

* async component error

* checkpt

* edit invoice mode

* integrate invoice generation

* invoices router

* scorched earth update invoices

* invoice helpers

* invoice helpers and sql

* ts ignore hopefully

+++ platform/flowglad-next/src/utils/purchaseSessionState.ts
@@ -6,10 +6,15 @@ import {
   updatePurchaseSession,
 } from '@/db/tableMethods/purchaseSessionMethods'
 import {
+  createPaymentIntentForInvoicePurchaseSession,
   createPaymentIntentForPurchaseSession,
   createSetupIntentForPurchaseSession,
 } from '@/utils/stripe'
-import { PriceType, PurchaseSessionStatus } from '@/types'
+import {
+  PriceType,
+  PurchaseSessionStatus,
+  PurchaseSessionType,
+} from '@/types'
 import { DbTransaction } from '@/db/types'
 import { PurchaseSession } from '@/db/schema/purchaseSessions'
 import { selectProductById } from '@/db/tableMethods/productMethods'
@@ -18,15 +23,26 @@ import { Purchase } from '@/db/schema/purchases'
 
 import { z } from 'zod'
 import { idInputSchema } from '@/db/tableUtils'
+import core from './core'
+import { Invoice } from '@/db/schema/invoices'
+import { InvoiceLineItem } from '@/db/schema/invoiceLineItems'
+import { FeeCalculation } from '@/db/schema/feeCalculations'
+import { selectCustomerProfileById } from '@/db/tableMethods/customerProfileMethods'
 
 const productPurchaseSessionCookieNameParamsSchema = z.object({
+  type: z.literal('product'),
   productId: z.string(),
 })
 
 const purchasePurchaseSessionCookieNameParamsSchema = z.object({
+  type: z.literal('purchase'),
   purchaseId: z.string(),
 })
 
+const invoicePurchaseSessionCookieNameParamsSchema = z.object({
+  type: z.literal('invoice'),
+  invoiceId: z.string(),
+})
 /**
  * SUBTLE CODE ALERT:
  * The order of z.union matters here!
@@ -40,10 +56,12 @@ const purchasePurchaseSessionCookieNameParamsSchema = z.object({
  * We actually want this because open purchases are more strict versions than variants
  *
  */
-export const purchaseSessionCookieNameParamsSchema = z.union([
-  purchasePurchaseSessionCookieNameParamsSchema,
-  productPurchaseSessionCookieNameParamsSchema,
-])
+export const purchaseSessionCookieNameParamsSchema =
+  z.discriminatedUnion('type', [
+    purchasePurchaseSessionCookieNameParamsSchema,
+    productPurchaseSessionCookieNameParamsSchema,
+    invoicePurchaseSessionCookieNameParamsSchema,
+  ])
 
 export const setPurchaseSessionCookieParamsSchema = idInputSchema.and(
   purchaseSessionCookieNameParamsSchema
@@ -63,10 +81,19 @@ export type PurchaseSessionCookieNameParams = z.infer<
 
 const purchaseSessionName = (
   params: PurchaseSessionCookieNameParams
-) =>
-  `purchase-session-id-${
-    'purchaseId' in params ? params.purchaseId : params.productId
-  }`
+) => {
+  const base = 'purchase-session-id-'
+  switch (params.type) {
+    case PurchaseSessionType.Product:
+      return base + params.productId
+    case PurchaseSessionType.Purchase:
+      return base + params.purchaseId
+    case PurchaseSessionType.Invoice:
+      return base + params.invoiceId
+    default:
+      throw new Error('Invalid purchase session type: ' + params.type)
+  }
+}
 
 /**
  * We must support multiple purchase session cookies on the client,
@@ -104,7 +131,36 @@ export const findPurchaseSession = async (
   return sessions[0]
 }
 
-export const createPurchaseSession = async (
+export const findPurchasePurchaseSession = async (
+  purchaseId: string,
+  transaction: DbTransaction
+) => {
+  return findPurchaseSession(
+    { purchaseId, type: PurchaseSessionType.Purchase },
+    transaction
+  )
+}
+
+export const findProductPurchaseSession = async (
+  productId: string,
+  transaction: DbTransaction
+) => {
+  return findPurchaseSession(
+    { productId, type: PurchaseSessionType.Product },
+    transaction
+  )
+}
+export const findInvoicePurchaseSession = async (
+  invoiceId: string,
+  transaction: DbTransaction
+) => {
+  return findPurchaseSession(
+    { invoiceId, type: PurchaseSessionType.Invoice },
+    transaction
+  )
+}
+
+export const createNonInvoicePurchaseSession = async (
   {
     variant,
     purchase,
@@ -116,15 +172,30 @@ export const createPurchaseSession = async (
   },
   transaction: DbTransaction
 ) => {
+  const purchaseSessionInsertCore = {
+    VariantId: variant.id,
+    status: PurchaseSessionStatus.Open,
+    expires: new Date(Date.now() + 1000 * 60 * 60 * 24), // 24 hours
+    OrganizationId,
+    livemode: variant.livemode,
+    ProductId: variant.ProductId,
+  } as const
+
+  const purchaseSesionInsert: PurchaseSession.Insert = purchase
+    ? {
+        ...purchaseSessionInsertCore,
+        PurchaseId: purchase.id,
+        InvoiceId: null,
+        type: PurchaseSessionType.Purchase,
+      }
+    : {
+        ...purchaseSessionInsertCore,
+        InvoiceId: null,
+        type: PurchaseSessionType.Product,
+      }
+
   const purchaseSession = await insertPurchaseSession(
-    {
-      VariantId: variant.id,
-      status: PurchaseSessionStatus.Open,
-      expires: new Date(Date.now() + 1000 * 60 * 60 * 24), // 24 hours
-      OrganizationId,
-      PurchaseId: purchase?.id,
-      livemode: variant.livemode,
-    },
+    purchaseSesionInsert,
     transaction
   )
   const organization = await selectOrganizationById(
@@ -165,9 +236,10 @@ export const createPurchaseSession = async (
       stripePaymentIntentId = paymentIntent.id
     }
   }
+
   const updatedPurchaseSession = await updatePurchaseSession(
     {
-      id: purchaseSession.id,
+      ...purchaseSession,
       stripePaymentIntentId,
       stripeSetupIntentId,
     },
@@ -175,6 +247,125 @@ export const createPurchaseSession = async (
   )
 
   return updatedPurchaseSession
+}
+
+export const findOrCreatePurchaseSession = async (
+  {
+    ProductId,
+    OrganizationId,
+    variant,
+    purchase,
+    type,
+  }: {
+    ProductId: string
+    OrganizationId: string
+    variant: Variant.Record
+    purchase?: Purchase.Record
+    type: PurchaseSessionType.Product | PurchaseSessionType.Purchase
+  },
+  transaction: DbTransaction
+) => {
+  const purchaseSession = await findPurchaseSession(
+    {
+      productId: ProductId,
+      purchaseId: purchase?.id,
+      type,
+    } as PurchaseSessionCookieNameParams,
+    transaction
+  )
+  if (
+    core.isNil(purchaseSession) ||
+    purchaseSession.VariantId !== variant.id
+  ) {
+    return createNonInvoicePurchaseSession(
+      { variant, OrganizationId, purchase },
+      transaction
+    )
+  }
+  return purchaseSession
+}
+
+const createInvoicePurchaseSession = async (
+  {
+    invoice,
+    invoiceLineItems,
+    feeCalculation,
+  }: {
+    invoice: Invoice.Record
+    invoiceLineItems: InvoiceLineItem.Record[]
+    feeCalculation?: FeeCalculation.Record
+  },
+  transaction: DbTransaction
+) => {
+  const customerProfile = await selectCustomerProfileById(
+    invoice.CustomerProfileId,
+    transaction
+  )
+  const purchaseSession = await insertPurchaseSession(
+    {
+      status: PurchaseSessionStatus.Open,
+      type: PurchaseSessionType.Invoice,
+      InvoiceId: invoice.id,
+      OrganizationId: invoice.OrganizationId,
+      CustomerProfileId: invoice.CustomerProfileId,
+      customerEmail: customerProfile.email,
+      customerName: customerProfile.name,
+      livemode: invoice.livemode,
+      PurchaseId: null,
+      VariantId: null,
+    },
+    transaction
+  )
+  const organization = await selectOrganizationById(
+    invoice.OrganizationId,
+    transaction
+  )
+  const paymentIntent =
+    await createPaymentIntentForInvoicePurchaseSession({
+      invoice,
+      organization,
+      purchaseSession,
+      invoiceLineItems: invoiceLineItems,
+      feeCalculation: feeCalculation,
+      stripeCustomerId: customerProfile.stripeCustomerId!,
+    })
+  const updatedPurchaseSession = await updatePurchaseSession(
+    {
+      ...purchaseSession,
+      stripePaymentIntentId: paymentIntent.id,
+    },
+    transaction
+  )
+  return updatedPurchaseSession
+}
+
+export const findOrCreateInvoicePurchaseSession = async (
+  {
+    invoice,
+    invoiceLineItems,
+    feeCalculation,
+  }: {
+    invoice: Invoice.Record
+    invoiceLineItems: InvoiceLineItem.Record[]
+    feeCalculation?: FeeCalculation.Record
+  },
+  transaction: DbTransaction
+) => {
+  const purchaseSession = await findPurchaseSession(
+    {
+      invoiceId: invoice.id,
+      type: PurchaseSessionType.Invoice,
+    },
+    transaction
+  )
+  if (purchaseSession) {
+    return purchaseSession
+  }
+
+  return createInvoicePurchaseSession(
+    { invoice, invoiceLineItems, feeCalculation },
+    transaction
+  )
 }
 
 type SetPurchaseSessionCookieParams = {
@@ -195,21 +386,34 @@ export const setPurchaseSessionCookie = async (
  * This strategy ensures we delete variant id
  * @param params
  */
-export const deletePurchaseSessionCookie = async (
-  params: PurchaseSessionCookieNameParams
-) => {
+export const deletePurchaseSessionCookie = async (params: {
+  productId?: string
+  purchaseId?: string
+  invoiceId?: string
+}) => {
   const cookieStore = await cookies()
-  if ('productId' in params) {
+  if ('productId' in params && params.productId) {
     await cookieStore.delete(
       purchaseSessionName({
         productId: params.productId,
+        type: PurchaseSessionType.Product,
       })
     )
   }
-  if ('purchaseId' in params) {
+  if ('purchaseId' in params && params.purchaseId) {
     await cookieStore.delete(
       purchaseSessionName({
         purchaseId: params.purchaseId,
+        type: PurchaseSessionType.Purchase,
+      })
+    )
+  }
+
+  if ('invoiceId' in params && params.invoiceId) {
+    await cookieStore.delete(
+      purchaseSessionName({
+        invoiceId: params.invoiceId,
+        type: PurchaseSessionType.Invoice,
       })
     )
   }
